Split reducers to each components
