+++ src/App-reducer.js
@@ -1,0 +1,20 @@
+import { UPDATE_SELECTED_STOCK, RESET } from "./store/constants";
+
+const initialState = {
+  selectedStock: ""
+};
+
+export const stockReducer = (state = initialState, { type, payload }) => {
+  switch (type) {
+    case UPDATE_SELECTED_STOCK:
+      return {
+        ...state,
+        selectedStock: payload
+      };
+    case RESET:
+      return { ...initialState, selectedStock: state.selectedStock };
+
+    default:
+      return state;
+  }
+};
