+++ src/features/search/components/Search.js
@@ -1,14 +1,13 @@
 import React, { useState, useRef, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { updateSearchQueryAction } from "../../../store/actions";
+import { updateSearchQueryAction } from "../../headline/redux/actions";
 import "./Search.css";
 
 export const Search = ({ updateStock }) => {
   const dispatch = useDispatch();
-  const filteredSymbols = useSelector(state => {
-    // console.log(state)
-    return state.stockData.selectedCompanySymbols;
-  });
+  const filteredSymbols = useSelector(
+    state => state.headlineData.selectedCompanySymbols
+  );
   const [isOpen, toggleIsOpen] = useState(false);
   const [stock, setStock] = useState("");
   const [symbol, setSymbol] = useState("");
